% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swper_verify.R
\name{swper_verify}
\alias{swper_verify}
\title{SWPER Verification Function}
\usage{
swper_verify(df, cols)
}
\arguments{
\item{df}{A data frame with the variables to be verified.}

\item{cols}{A vector with the names or indices of the columns to be verified.}
}
\value{
A data frame with the results of the verification, including the variable name, item name, valid codes, invalid values, and missing values for each specified column.
}
\description{
SWPER Verification Function
}
\details{
This function checks that the variables in a given data frame are correctly coded to construct the SWPER index score. Users can specify which columns to use and which codes correspond to each response option.
}
