% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swper_score.R
\name{swper_score}
\alias{swper_score}
\title{Calculate Survey-based Women emPowERment index (SWPER)}
\usage{
swper_score(df, cols, indices, estandarizacion = NULL)
}
\arguments{
\item{df}{The dataset to use.}

\item{cols}{The columns of the dataset to use.}

\item{indices}{The indices to use in the SWPER index calculation.}

\item{estandarizacion}{The region of standardization to use.}
}
\value{
A data frame containing the SWPER index for each combination of indices and standardization region.
}
\description{
This function calculates the SWPER score for a given dataset and set of indices, and returns the result as a data frame.
}
